Remove version verification
Separately, the LMS oauth_dispatch is being updated to be capable of
returning JWTs with different major versions when specifically
requested.

For the planned v1.x to v2.x JWT, the edx-drf-extensions _verify_version
call didn't actually help, and just forced an upgrade/deployment of
every service. Therefore, this check is being removed until/unless it
can be coded to solve a particular problem.

ARCH-379