fix for edx-drf-extensions 3.0.0
In 3.0.0, the switch `oauth2.enforce_jwt_scopes` was removed, which
starts checking is_restricted in JWTs and works fine for JWTs created
with the LMS, but was broken for ecommerce service JWTs which
were meant to be decoded with its own jwt_decode_handler.

This fix updates the JWT code to respect the JWT_DECODE_HANDLER setting
in JWT_AUTH, and uses the configured handler rather than assuming the
edx-drf-extensions version will always be used.

ARCHBOM-1036