refactor: remove check for non-jwt tokens
At one point, drf-jwt had an issue introduced where it was raising an
error instead of returning None for non-jwt tokens, for example, bearer
tokens. A check was temporarily added to JwtAuthentication to ensure
that we continued to return None.

Since then, our team got a more permanent fix merged upstream in:
https://github.com/edx/edx-drf-extensions/pull/136

We will keep the unit test added with the original fix to ensure that
this continues to work.