test: Update tests to also test forgiving JWT Auth.
Run all the existing tests with both forgiving and original JWT Auth
behaviors.  Only one test needed to be modified.  Previously we were
raising an exception when CSRF checks failed within the authentication
process but with forgiving auth, we no longer rais an exception so we
update the one test to handle both cases for now.

This will get cleaned when we hopefully move forward with only having
the forgiving JWT auth flow in the future.