test: Add tests for the redirect middleware.
Previously, it would result in 401s for endpoints that did not have the
relevant middleware and permissions classes to auto turn on
USE_JWT_COOKIE_HEADER.  But with the forgiving JWT auth, as long as the
cookie is set, the auth will work.  This will mean more cases where
hitting endpoists will be successful for users that have the JWT cookies
set.