fix: Add more monitoring to the JWT Auth Class.
The default exception that comes from the JSONWebTokenAuthentication
does not have enough information about why the JWT auth failed in the
exception that it propagates.  However, we can get this information by
looking at the PyJWT exception that was thrown so we added some tooling
to get that lower level exception if it exists.